+++ code/cell.modules/pkg.sys/net/src/web.PeerNetwork/PeerNetwork.TEST.ts
@@ -1,5 +1,5 @@
-import { expect, Test, cuid, rx, TEST } from '../test';
 import { PeerNetwork } from '.';
+import { cuid, expect, rx, TEST, Test } from '../web.test';
 
 const signal = TEST.SIGNAL;
